# docs/swagger.yaml
definitions:
  api.Error:
    properties:
      error:
        type: string
    type: object
  api.Job:
    properties:
      args:
        items:
          type: string
        type: array
      customer_id:
        type: string
      end_time:
        $ref: '#/definitions/timestamppb.Timestamp'
      error:
        type: string
      id:
        type: string
      input_id:
        type: string
      output_id:
        type: string
      start_time:
        $ref: '#/definitions/timestamppb.Timestamp'
      status:
        type: string
      task_type:
        type: string
    type: object
  rototiller.Storage:
    properties:
      create_time:
        $ref: '#/definitions/timestamppb.Timestamp'
      customer_id:
        type: string
      id:
        type: string
      last_used:
        $ref: '#/definitions/timestamppb.Timestamp'
      name:
        type: string
      status:
        type: string
    type: object
  rototiller.Task:
    properties:
      kind:
        type: string
      params:
        items:
          type: string
        type: array
      type:
        type: string
    type: object
  timestamppb.Timestamp:
    properties:
      nanos:
        description: |-
          Non-negative fractions of a second at nanosecond resolution. Negative
          second values with fractions must still have non-negative nanos values
          that count forward in time. Must be from 0 to 999,999,999
          inclusive.
        type: integer
      seconds:
        description: |-
          Represents seconds of UTC time since Unix epoch
          1970-01-01T00:00:00Z. Must be from 0001-01-01T00:00:00Z to
          9999-12-31T23:59:59Z inclusive.
        type: integer
    type: object
info:
  contact:
    email: logsquaredn@gmail.com
    name: logsquaredn
    url: https://rototiller.logsquaredn.io/
  license:
    name: logsquaredn
paths:
...
  /api/v1/jobs:
    get:
      description: |-
        Get a list of jobs based on API Key

        &emsp; - API Key is required either as a query parameter or a header
      parameters:
      - description: API Key header
        in: header
        name: X-API-Key
        type: string
      - description: API Key query parameter
        in: query
        name: api-key
        type: string
      - description: Offset of jobs to return
        in: query
        name: offset
        type: integer
      - description: Limit of jobs to return
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.Job'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      summary: Get a list of jobs
      tags:
      - Job
  /api/v1/jobs/{id}:
    get:
      description: |-
        Get the metadata of a job. This can be used as a way to track job status

        &emsp; - API Key is required either as a query parameter or a header
      parameters:
      - description: API Key query parameter
        in: query
        name: api-key
        type: string
      - description: API Key header
        in: header
        name: X-API-Key
        type: string
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Job'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      summary: Get a job
      tags:
      - Job
...
  /api/v1/jobs/buffer:
    post:
      consumes:
      - application/json
      - ' application/zip'
      description: |-
        <b><u>Create a buffer job</u></b>
        &emsp; - Buffers every geometry by the given distance

        &emsp; - For extra info: https://gdal.org/api/vector_c_api.html#_CPPv412OGR_G_Buffer12OGRGeometryHdi
        &emsp; - API Key is required either as a query parameter or a header
        &emsp; - Pass the geospatial data to be processed in the request body OR
        &emsp; - Pass the ID of an existing dataset with an empty request body
        &emsp; - This task accepts a ZIP containing a shapefile or GeoJSON input
        &emsp; - This task will automatically generate both GeoJSON and ZIP (shapfile) output
      parameters:
      - description: API Key query parameter
        in: query
        name: api-key
        type: string
      - description: Required if passing geospatial data in request body
        in: header
        name: Content-Type
        type: string
      - description: API Key header
        in: header
        name: X-API-Key
        type: string
      - description: ID of existing dataset to use
        in: query
        name: input
        type: string
      - description: ID of existing job whose input dataset to use
        in: query
        name: input-of
        type: string
      - description: ID of existing job whose output dataset to use
        in: query
        name: output-of
        type: string
      - description: Buffer distance
        in: query
        name: buffer-distance
        required: true
        type: integer
      - description: Quadrant Segment count
        in: query
        name: quadrant-segment-count
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Job'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      summary: Create a buffer job
      tags:
      - Job
  /api/v1/jobs/filter:
    post:
      consumes:
      - application/json
      - ' application/zip'
      description: |-
        <b><u>Create a filter job</u></b>
        &emsp; - Drops features and their geometries that don't match the given filter

        &emsp; - API Key is required either as a query parameter or a header
        &emsp; - Pass the geospatial data to be processed in the request body OR
        &emsp; - Pass the ID of an existing dataset with an empty request body
        &emsp; - This task accepts a ZIP containing a shapefile or GeoJSON input
        &emsp; - This task will automatically generate both GeoJSON and ZIP (shapfile) output
      parameters:
      - description: API Key query parameter
        in: query
        name: api-key
        type: string
      - description: Required if passing geospatial data in request body
        in: header
        name: Content-Type
        type: string
      - description: API Key header
        in: header
        name: X-API-Key
        type: string
      - description: ID of existing dataset to use
        in: query
        name: input
        type: string
      - description: ID of existing job whose input dataset to use
        in: query
        name: input-of
        type: string
      - description: ID of existing job whose output dataset to use
        in: query
        name: output-of
        type: string
      - description: Column to filter on
        in: query
        name: filter-column
        required: true
        type: string
      - description: Value to filter on
        in: query
        name: filter-value
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Job'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      summary: Create a filter job
      tags:
      - Job
  /api/v1/jobs/rasterlookup:
    post:
      consumes:
      - application/json
      - ' application/zip'
      description: |-
        <b><u>Create a raster lookup job</u></b>
        &emsp; - Returns the value of each requested band of which the given point intersects

        &emsp; - API Key is required either as a query parameter or a header
        &emsp; - Pass the geospatial data to be processed in the request body OR
        &emsp; - Pass the ID of an existing dataset with an empty request body
        &emsp; - This task accepts a ZIP containing a single TIF file. Valid extensions are: tif, tiff, geotif, geotiff
        &emsp; - This task will generate JSON output
      parameters:
      - description: API Key query parameter
        in: query
        name: api-key
        type: string
      - description: Required if passing geospatial data in request body
        in: header
        name: Content-Type
        type: string
      - description: API Key header
        in: header
        name: X-API-Key
        type: string
      - description: ID of existing dataset to use
        in: query
        name: input
        type: string
      - description: ID of existing job whose input dataset to use
        in: query
        name: input-of
        type: string
      - description: ID of existing job whose output dataset to use
        in: query
        name: output-of
        type: string
      - description: Comma separated list of bands
        in: query
        name: bands
        required: true
        type: string
      - description: Longitude
        in: query
        name: longitude
        required: true
        type: number
      - description: Latitude
        in: query
        name: latitude
        required: true
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Job'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      summary: Create a raster lookup job
      tags:
      - Job
  /api/v1/jobs/removebadgeometry:
    post:
      consumes:
      - application/json
      - ' application/zip'
      description: |-
        <b><u>Create a remove bad geometry job</u></b>
        &emsp; - Drops geometries that are invalid

        &emsp; - For extra info: https://gdal.org/api/vector_c_api.html#_CPPv413OGR_G_IsValid12OGRGeometryH
        &emsp; - API Key is required either as a query parameter or a header
        &emsp; - Pass the geospatial data to be processed in the request body OR
        &emsp; - Pass the ID of an existing dataset with an empty request body
        &emsp; - This task accepts a ZIP containing a shapefile or GeoJSON input
        &emsp; - This task will automatically generate both GeoJSON and ZIP (shapfile) output
      parameters:
      - description: API Key query parameter
        in: query
        name: api-key
        type: string
      - description: Required if passing geospatial data in request body
        in: header
        name: Content-Type
        type: string
      - description: API Key header
        in: header
        name: X-API-Key
        type: string
      - description: ID of existing dataset to use
        in: query
        name: input
        type: string
      - description: ID of existing job whose input dataset to use
        in: query
        name: input-of
        type: string
      - description: ID of existing job whose output dataset to use
        in: query
        name: output-of
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Job'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      summary: Create a remove bad geometry job
      tags:
      - Job
  /api/v1/jobs/reproject:
    post:
      consumes:
      - application/json
      - ' application/zip'
      description: |-
        <b><u>Create a reproject job</u></b>
        &emsp; - Reprojects all geometries to the given projection

        &emsp; - API Key is required either as a query parameter or a header
        &emsp; - Pass the geospatial data to be processed in the request body OR
        &emsp; - Pass the ID of an existing dataset with an empty request body
        &emsp; - This task accepts a ZIP containing a shapefile or GeoJSON input
        &emsp; - This task will automatically generate both GeoJSON and ZIP (shapfile) output
      parameters:
      - description: API Key query parameter
        in: query
        name: api-key
        type: string
      - description: Required if passing geospatial data in request body
        in: header
        name: Content-Type
        type: string
      - description: API Key header
        in: header
        name: X-API-Key
        type: string
      - description: ID of existing dataset to use
        in: query
        name: input
        type: string
      - description: ID of existing job whose input dataset to use
        in: query
        name: input-of
        type: string
      - description: ID of existing job whose output dataset to use
        in: query
        name: output-of
        type: string
      - description: Target projection EPSG
        in: query
        name: target-projection
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Job'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      summary: Create a reproject job
      tags:
      - Job
  /api/v1/jobs/vectorlookup:
    post:
      consumes:
      - application/json
      - ' application/zip'
      description: |-
        <b><u>Create a vector lookup job</u></b>
        &emsp; - Returns the feature and geometry of which the given point intersects

        &emsp; - API Key is required either as a query parameter or a header
        &emsp; - Pass the geospatial data to be processed in the request body OR
        &emsp; - Pass the ID of an existing dataset with an empty request body
        &emsp; - This task accepts a ZIP containing a shapefile or GeoJSON input
        &emsp; - This task will automatically generate both GeoJSON and ZIP (shapfile) output
      parameters:
      - description: API Key query parameter
        in: query
        name: api-key
        type: string
      - description: Required if passing geospatial data in request body
        in: header
        name: Content-Type
        type: string
      - description: API Key header
        in: header
        name: X-API-Key
        type: string
      - description: ID of existing dataset to use
        in: query
        name: input
        type: string
      - description: ID of existing job whose input dataset to use
        in: query
        name: input-of
        type: string
      - description: ID of existing job whose output dataset to use
        in: query
        name: output-of
        type: string
      - description: Comma separated list of attributes
        in: query
        name: attributes
        required: true
        type: string
      - description: Longitude
        in: query
        name: longitude
        required: true
        type: number
      - description: Latitude
        in: query
        name: latitude
        required: true
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Job'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      summary: Create a vector lookup job
      tags:
      - Job
  /api/v1/storages:
    get:
      description: |-
        Get a list of stored datasets based on API Key

        &emsp; - API Key is required either as a query parameter or a header
      parameters:
      - description: API Key header
        in: header
        name: X-API-Key
        type: string
      - description: API Key query parameter
        in: query
        name: api-key
        type: string
      - description: Offset of storages to return
        in: query
        name: offset
        type: integer
      - description: Limit of storages to return
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/rototiller.Storage'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      summary: Get a list of storage
      tags:
      - Storage
    post:
      consumes:
      - application/json
      - ' application/zip'
      description: |-
        Stores a dataset. The ID of this stored dataset can be used as input to jobs

        &emsp; - API Key is required either as a query parameter or a header
        &emsp; - Pass the geospatial data to be stored in the request body
      parameters:
      - description: API Key query parameter
        in: query
        name: api-key
        type: string
      - description: API Key header
        in: header
        name: X-API-Key
        type: string
      - description: Storage name
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rototiller.Storage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      summary: Create a storage
      tags:
      - Storage
  /api/v1/storages/{id}:
    get:
      description: |-
        Get the metadata of a stored dataset

        &emsp; - API Key is required either as a query parameter or a header
      parameters:
      - description: API Key query parameter
        in: query
        name: api-key
        type: string
      - description: API Key header
        in: header
        name: X-API-Key
        type: string
      - description: Storage ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rototiller.Storage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      summary: Get a storage
      tags:
      - Storage
...
  /api/v1/tasks:
    get:
      description: '&emsp; - API Key is required either as a query parameter or a
        header'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/rototiller.Task'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      summary: Get a list of task types
      tags:
      - Task
  /api/v1/tasks/{type}:
    get:
      description: '&emsp; - API Key is required either as a query parameter or a
        header'
      parameters:
      - description: Task type
        in: path
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rototiller.Task'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      summary: Get a task type
      tags:
      - Task
swagger: "2.0"