# types.go
package rototiller

import "github.com/logsquaredn/rototiller/pkg/api"

type Customer = api.Customer

type Event = api.Event

type Job = api.Job

type Storage = api.Storage

type Task = api.Task

type TaskKind = api.TaskKind

type TaskType = api.TaskType

type StorageStatus = api.StorageStatus